Update mute status of self video when not floating as well (#2111)
